pyproject.toml
[project]
name = "studio-fury"
description = "Suite profesional para dirección de arte y flujos de trabajo avanzados en ComfyUI."
version = "2.0.0"
license = { file = "LICENSE" }
dependencies = ["torch", "pillow", "numpy", "aiohttp"]

[project.urls]
Repository = "https://github.com/FuryNocturn/ComfyUI-Studio-Fury"
Documentation = "https://github.com/FuryNocturn/ComfyUI-Studio-Fury/wiki"
"Bug Tracker" = "https://github.com/FuryNocturn/ComfyUI-Studio-Fury/issues"

[tool.comfy]
PublisherId = "furynocturntv"
DisplayName = "StudioFury"
Icon = "https://github.com/FuryNocturn/ComfyUI-Studio-Fury/blob/master/ico/ICO.png"
Banner = "https://github.com/FuryNocturn/ComfyUI-Studio-Fury/blob/master/ico/Banner.png"